Rename files so they are under the absurd 100-byte filename limit.
