+++ src/mlpack/methods/decision_tree/select_functions/select_functions.hpp
@@ -13,7 +13,7 @@
 #define MLPACK_METHODS_DECISION_TREE_SELECT_FUNCTIONS_SELECT_FUNCTIONS_HPP
 
 #include "all_dimension_select.hpp"
-#include "multiple_random_dimension_select.hpp"
+#include "mult_random_dimension_select.hpp"
 #include "random_dimension_select.hpp"
 
 #endif
