+++ src/mlpack/methods/decision_tree/splits/best_binary_categorical_split.hpp
@@ -1,5 +1,5 @@
 /**
- * @file methods/decision_tree/best_binary_categorical_split.hpp
+ * @file methods/decision_tree/splits/best_binary_categorical_split.hpp
  * @author Nikolay Apanasov (nikolay@apanasov.org)
  *
  * A tree splitter that finds the best binary categorical split.
@@ -9,8 +9,8 @@
  * 3-clause BSD license along with mlpack.  If not, see
  * http://www.opensource.org/licenses/BSD-3-Clause for more information.
  */
-#ifndef MLPACK_METHODS_DECISION_TREE_BEST_BINARY_CATEGORICAL_SPLIT_HPP
-#define MLPACK_METHODS_DECISION_TREE_BEST_BINARY_CATEGORICAL_SPLIT_HPP
+#ifndef MLPACK_METHODS_DECISION_TREE_SPLITS_BEST_BINARY_CATEGORICAL_SPLIT_HPP
+#define MLPACK_METHODS_DECISION_TREE_SPLITS_BEST_BINARY_CATEGORICAL_SPLIT_HPP
 
 #include <mlpack/prereqs.hpp>
