+++ src/mlpack/methods/decision_tree/splits/splits.hpp
@@ -1,5 +1,5 @@
 /**
- * @file methods/decision_tree/split_functions/split_functions.hpp
+ * @file methods/decision_tree/split_functions/splits.hpp
  * @author Abhimanyu Dayal
  *
  * Include all split policies for decision trees.
@@ -9,8 +9,8 @@
  * 3-clause BSD license along with mlpack.  If not, see
  * http://www.opensource.org/licenses/BSD-3-Clause for more information.
  */
-#ifndef MLPACK_METHODS_DECISION_TREE_SPLIT_FUNCTIONS_SPLIT_FUNCTIONS_HPP
-#define MLPACK_METHODS_DECISION_TREE_SPLIT_FUNCTIONS_SPLIT_FUNCTIONS_HPP
+#ifndef MLPACK_METHODS_DECISION_TREE_SPLIT_FUNCTIONS_SPLITS_HPP
+#define MLPACK_METHODS_DECISION_TREE_SPLIT_FUNCTIONS_SPLITS_HPP
 
 #include "all_categorical_split.hpp"
 #include "best_binary_numeric_split.hpp"
